+++ Engine/src/ECS/CoreComponents/Material.hpp
@@ -12,7 +12,7 @@ struct Material
     float roughness  = 1.0f;
     float metallic   = 0.0;
 
-    uint32_t _ubSlot;       // internal
+    uint64_t _ubSlot;       // internal
     uint32_t _textureSlot;  // internal
 };
